python/BigQuerify/requirements.txt
json>=2.0.9
google-cloud-bigquery>=1.6.0
